WIP:refactor: migrate to standalone AI agent service
- Remove old agent control implementation
- Add dedicated tt_studio_agent service in docker-compose
- Update frontend to always show AI agent selector
- Modify backend to use standalone agent service
- Update documentation references
