+++ app/agent/.dockerignore
@@ -1,0 +1,22 @@
+__pycache__
+*.pyc
+*.pyo
+*.pyd
+.Python
+env
+venv
+.venv
+pip-log.txt
+.git
+.gitignore
+README.md
+*.png
+*.jpg
+*.jpeg
+*.gif
+.DS_Store
+.env
+.env.local
+.env.development
+.env.test
+.env.production
