+++ app/backend/docker_control/views.py
@@ -19,7 +19,6 @@ import os
 from .forms import DockerForm
 from .docker_utils import (
     run_container,
-    run_agent_container, 
     stop_container,
     get_container_status,
     perform_reset,
@@ -61,18 +60,7 @@ class StopView(APIView):
             container_id = request.data.get("container_id")
             logger.info(f"Received request to stop container with ID: {container_id}")
 
-            # Find agent container 
-            container_name = client.containers.get(container_id).name
-            last_underscore_index = container_name.rfind('_')
-            llm_host_port = container_name[last_underscore_index + 1:]
-
-            agent_container_name = f"ai_agent_container_{llm_host_port}"
-            all_containers = client.containers.list(all=True)
-            for container in all_containers:
-                if container.name == agent_container_name: # if the agent corresponding agent container is found
-                    stop_container(container.id) # remove the agent container 
-
-            # Stop the container
+            # Stop the main container
             stop_response = stop_container(container_id)
             logger.info(f"Stop response: {stop_response}")
 
@@ -173,6 +161,9 @@ class StatusView(APIView):
         return Response(data, status=status.HTTP_200_OK)
 
 
+
+
+
 class DeployView(APIView):
     def post(self, request, *args, **kwargs):
         serializer = DeploymentSerializer(data=request.data)
@@ -181,12 +172,7 @@ class DeployView(APIView):
             weights_id = request.data.get("weights_id")
             impl = model_implmentations[impl_id]
             response = run_container(impl, weights_id)
-            if os.getenv("TAVILY_API_KEY") == "your-tavily-api-key":
-                agent_api_key_set = False 
-            else:
-                agent_api_key_set = True
-            if impl.model_type == ModelTypes.CHAT and agent_api_key_set:
-                run_agent_container(response["container_name"], response["port_bindings"], impl) # run agent container that maps to appropriate LLM container
+            
             return Response(response, status=status.HTTP_201_CREATED)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
