+++ app/backend/model_control/views.py
@@ -104,26 +104,30 @@ class InferenceView(APIView):
 class AgentView(APIView):
     def post(self, request, *agrs, **kwargs):
         logger.info(f"URL '/agent/' accessed via POST method by {request.META['REMOTE_ADDR']}")        
-        data = request.data 
+        data = request.data.copy()  # Make a copy to avoid modifying the original
         logger.info(f"AgentView data:={data}")
-        serializer = InferenceSerializer(data=data)
-        if serializer.is_valid():
-            deploy_id = data.pop("deploy_id")
-            logger.info(f"Deploy ID: {deploy_id}")
-            deploy = get_deploy_cache()[deploy_id]
-            colon_idx = deploy["internal_url"].rfind(":")
-            underscore_idx = deploy["internal_url"].rfind("_")
-            llm_host_port = deploy["internal_url"][underscore_idx + 2: colon_idx] # add 2 to remove the p
-            # agent port on host is 200 + the llm host port
-            internal_url = f"http://ai_agent_container_p{llm_host_port}:{int(llm_host_port) + 200}/poll_requests"
-            logger.info(f"internal_url:= {internal_url}")
-            logger.info(f"using vllm model:= {deploy["model_impl"].model_name}")
+        
+        # For agent requests, we don't need to validate deploy_id since agents can work with cloud models
+        deploy_id = data.get("deploy_id", "")
+        logger.info(f"Deploy ID: {deploy_id}")
+        
+        # Check if we have a valid deployment, if not, proceed without it (for cloud/agent mode)
+        deploy_cache = get_deploy_cache()
+        if deploy_id and deploy_id in deploy_cache:
+            deploy = deploy_cache[deploy_id]
+            logger.info(f"using vllm model:= {deploy['model_impl'].model_name}")
             data["model"] = deploy["model_impl"].hf_model_id
-            logger.info(f"Using internal url: {internal_url}")
-            response_stream = stream_response_from_agent_api(internal_url, data)
-            return StreamingHttpResponse(response_stream, content_type="text/plain")
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            logger.info("No valid deployment found, proceeding with agent-only mode (cloud LLM)")
+            # Remove deploy_id from data since it's not needed for agent
+            data.pop("deploy_id", None)
+        
+        # Use the standalone agent service instead of dynamically created containers
+        internal_url = "http://tt_studio_agent:8080/poll_requests"
+        logger.info(f"internal_url:= {internal_url}")
+        logger.info(f"Using internal url: {internal_url}")
+        response_stream = stream_response_from_agent_api(internal_url, data)
+        return StreamingHttpResponse(response_stream, content_type="text/plain")
 
 class ModelHealthView(APIView):
     def get(self, request, *args, **kwargs):
